Update NumberTools with much more precise sin/cos
They're almost exactly the same speed as before, but they have
practically no difference from Math.sin() or Math.cos() in results
unless given very large inputs or you're looking extremely close for
differences (there should be less than 0.001 difference from Math's
result). Graph here https://www.desmos.com/calculator/g0ebg0fjmr .

+++ squidlib-util/src/main/java/squidpony/squidmath/IntDoubleOrderedMap.java
@@ -1688,7 +1688,7 @@ public class IntDoubleOrderedMap implements SortedMap<Integer, Double>, java.io.
          * Delegates to the type-specific <code>rem()</code> method.
          */
         public boolean remove(Object ok) {
-            return rem((Integer) ok);
+            return rem(ok);
         }
 
         /**
@@ -1702,7 +1702,7 @@ public class IntDoubleOrderedMap implements SortedMap<Integer, Double>, java.io.
          * Delegates to the corresponding type-specific method.
          */
         public boolean rem(final Object o) {
-            return rem((Integer) o);
+            return rem(o);
         }
 
         /**
