+++ squidlib-util/src/main/java/squidpony/squidmath/MummyNoise.java
@@ -69,7 +69,7 @@ public class MummyNoise implements Noise.Noise2D, Noise.Noise3D, Noise.Noise4D, 
     @Override
     public double getNoiseWithSeed(final double x, final double y, final long seed) {
         final long
-                rs = ThrustRNG.determine(seed ^ (long) ~seed << 32),
+                rs = ThrustRNG.determine(seed ^ ~seed << 32),
                 rx = 0x9E3779B97F4A7C15L * (rs >>> 23 ^ rs << 23) * (rs | 1L),
                 ry = 0x9E3779B97F4A7C15L * (rx >>> 23 ^ rx << 23) * (rx | 1L);
         return getNoiseWithSeeds(x, y, rx, ry);
@@ -100,7 +100,7 @@ public class MummyNoise implements Noise.Noise2D, Noise.Noise3D, Noise.Noise4D, 
     @Override
     public double getNoiseWithSeed(final double x, final double y, final double z, final long seed) {
         final long
-                rs = ThrustRNG.determine(seed ^ (long) ~seed << 32),
+                rs = ThrustRNG.determine(seed ^ ~seed << 32),
                 rx = 0x9E3779B97F4A7C15L * (rs >>> 23 ^ rs << 23) * (rs | 1L),
                 ry = 0x9E3779B97F4A7C15L * (rx >>> 23 ^ rx << 23) * (rx | 1L),
                 rz = 0x9E3779B97F4A7C15L * (ry >>> 23 ^ ry << 23) * (ry | 1L);
@@ -140,7 +140,7 @@ public class MummyNoise implements Noise.Noise2D, Noise.Noise3D, Noise.Noise4D, 
     @Override
     public double getNoiseWithSeed(final double x, final double y, final double z, final double w, final long seed) {
         final long
-                rs = ThrustRNG.determine(seed ^ (long) ~seed << 32),
+                rs = ThrustRNG.determine(seed ^ ~seed << 32),
                 rx = 0x9E3779B97F4A7C15L * (rs >>> 23 ^ rs << 23) * (rs | 1L),
                 ry = 0x9E3779B97F4A7C15L * (rx >>> 23 ^ rx << 23) * (rx | 1L),
                 rz = 0x9E3779B97F4A7C15L * (ry >>> 23 ^ ry << 23) * (ry | 1L),
@@ -197,7 +197,7 @@ public class MummyNoise implements Noise.Noise2D, Noise.Noise3D, Noise.Noise4D, 
     @Override
     public double getNoiseWithSeed(final double x, final double y, final double z, final double w, final double u, final double v, long seed) {
         final long
-                rs = ThrustRNG.determine(seed ^ (long) ~seed << 32),
+                rs = ThrustRNG.determine(seed ^ ~seed << 32),
                 rx = 0x9E3779B97F4A7C15L * (rs >>> 23 ^ rs << 23) * (rs | 1L),
                 ry = 0x9E3779B97F4A7C15L * (rx >>> 23 ^ rx << 23) * (rx | 1L),
                 rz = 0x9E3779B97F4A7C15L * (ry >>> 23 ^ ry << 23) * (ry | 1L),
